refactor common code to module
